Modification of Old Code for New FTC SDK 2.0
Essentially, modified all of last year's code so that it can compile and
run (untested as of now though) with the new SDK. In order to accomplish
this:

Removed usage of FTCRobotControllerActivity in Team OpModes.

Modified the gradle files to include the libraries for our navX-micro
IMU.

Changed package labels on all of our Team OpModes appropriately.

Added OpMode register annotations, and the required imports for the
annotations to work, to all Team OpModes.

Modified the line of code in OpMode_5220 which changes the I2C address
of one of the color sensors to use the new syntax for setting I2C
addresses.