Updates 9-12-2016
OPMODE_5220:

Made runConditions() final.

TELEOP:

Ported over everything for mecanum wheel drivetrain TeleOp control from
our old Cascade Effect code. Still need to test to see if everything is
correct, if or if things need to be switched around a bit.

Added use of Gamepad objects for storing previous states of the gamepad.
It is not yet tested. If using this method works, then the individual
booleans for the previous states of all the gamepad values can be
eliminated in favor of simply using these new prevGamepad objects.

OTHER:

Added new drivecontrol package. Inside shall be extremely simple
programs to drive, turn, and strafe until the program is stopped. These
programs should be listed by the OpModeRegister in a seperate folder, on
the driver station phone. Hopefully this will make manual drive control
without a gamepad easier. Right now only programs to drive forwards and
backwars are in the package. If they succeed, then more similar simple
programs will be added to that package.