Updates 9-13-2016
Added new methods in OpMode_5220 to allow for pre-programmed strafing in
the autonomous program. Currently, no gyro or encoder syncing for
strafing. Also, the distance conversion from inches to encoder counts
for strafing needs to be experimentally determined.

Deleted some old commented-out code in OpMode_5220.

Added a reminder (as a comment) in followLineUntilTouch() in Autonomous
to construct a new and improved line following algorithm using mecanum
wheels soon.