Updates 9-28-2016
Renamed a bunch of our code files, removing unnecessary version numbers
in name (since we use Git for version control)

Made copies of the old programs that compile but are not applicable to
this years robot. These copies are labeled with an underscore followed
by "Old" at the end. They are currently disabled in the OpModeRegister.
If we ever want to run last years robot with the new SDK, we can use
these "Old" files.