Updates 10-15-2016
Added shooter motor and shooter tilt servo.

Added simple TeleOp control for the shooter motor: if the A Button on
gamepad 1 is pressed down, the motor power is set to 100%. Otherwise it
is set to zero.

Added a whole bunch of simple drive control programs.

Added a simple strafing and moving test program to autonomous.

Eliminated a few problematic instances of using the IMU or other sensors
(that we currently don't have) in the code.