Updates 10-21-2016
Added door servo and sweeper (squirrel cheek) motor.

Added controls for door and sweeper to TeleOp.

Fixed shoot method to account for the NeveRest 60 having 1680 encooder
counts per rotation as opposed to the NeveReest 40's 1120. Also added
writeToLog stuff in the shoot method for the encode values

Added shoote encoder values to the debug display loop

Wrote loop to auto-load and shoot balls (the shootAll method) in OpMode,
and added it to the autonomous test method. It only uses time and the
shooter encoder right now, not any ball detector sensor. That sensor
will need to be added sometime soon.